separate google and gemini
